backend/requirements.txt
fastapi
uvicorn[standard]
openai
pydantic
python-dotenv
langchain
langchain-openai
